backend/requirements-dev.txt
pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
black==24.4.2
flake8==7.0.0
mypy==1.10.0
django-stubs==5.0.0
djangorestframework-stubs==3.15.0
